requirements.txt
pyteomics>=4.7.5
lxml>=5.2
numpy>=1.26
pandas>=2.2
matplotlib>=3.8
tqdm>=4.66
Pillow>=10.0
tifffile>=2024.1
openpyxl>=3.1
